cleanup: remove trend explorer
